+++ loops/loop-mediatheque.php
@@ -6,8 +6,11 @@
  * @subpackage CNRS_Web_Kit
  * @since CNRS Web Kit 1.0
  * 
- * Loop Name: Boucle d'actialités
+ * Loop Name: Boucle Mediathèque
  */
+
+// gNote get_rnd_image_size('mediatheque', $iteration_number) return 400 excepted for 1st modulo 9  image.
+
 ?>
 <article id="post-<?php echo $current_item->value('ID'); ?>" <?php post_class([$current_item->value('post_type'), 'size' . get_rnd_image_size('mediatheque', $iteration_number)], $current_item->value('ID')); ?>>
     <div class="chapoContainer">
@@ -26,15 +29,4 @@
             link: '<?php echo wp_get_attachment_image_src(get_post_thumbnail_id($current_item->value('ID')), 'full', false)[0]; ?>',
          }
     </script>
-    <?php
-    /*
-      <header class="entry-header">
-      <h1 class="entry-title"><a href="<?php echo get_permalink($current_item->value('ID')); ?>"><?php echo $current_item->value('post_title'); ?></a></h1>
-      </header>
-      <div class="entry-content">
-      <div><?php echo $current_item->value('chapo'); ?></div>
-      <div><?php echo $current_item->value('credits'); ?></div>
-      </div>
-     */
-    ?>
 </article><!-- #post-## -->