+++ sidebar-content-bottom.php
@@ -12,6 +12,12 @@ if (!is_active_sidebar('sidebar-2') && !is_active_sidebar('sidebar-3')) {
 
 // If we get this far, we have widgets. Let's do this.
 global $cnrs_global_params;
+
+
+if ($cnrs_global_params->field('newsletter_sur_la_page_daccueil')) {
+    // include for is_plugin_active which is an admin only function
+    include_once( ABSPATH . 'wp-admin/includes/plugin.php' );
+}
 if ($cnrs_global_params->field('newsletter_sur_la_page_daccueil') && is_plugin_active('newsletter') ) {
     ?>
     <!-- Start Actus Widget -->