+++ template-parts/bottom-evenement.php
@@ -18,7 +18,7 @@ $custom_params->limit = 2;
 $evenement_data = new CnrswebkitPageItemsList('evenement', $custom_params);
 $liste_evenement_url= get_permalink($cnrs_global_params->field('pageliste_evenement')['ID']);
 if ($liste_evenement_url) {
-	$liste_evenement_url = '<a href="'. $liste_evenement_url . '">' . __('Return to Events list', 'cnrswebkit') . '</a>';
+	$liste_evenement_url = '<a href="'. $liste_evenement_url . '">' . __('Browse Events list', 'cnrswebkit') . '</a>';
 }
 
 if ($evenement_data->has_items()) {